refactor: common types in lib
